feat: install redux and redux-toolkit apply ducks architecture
