+++ src/app/state/ducks/Card/index.tsx
@@ -1,0 +1,1 @@
+export { default as CardReducer } from "./reducers";
