+++ src/app/state/ducks/CardList/index.ts
@@ -1,0 +1,1 @@
+export {default as CardListReducer} from './reducers';
