+++ src/app/views/components/CardModal/components/NewComment/index.tsx
@@ -1,0 +1,1 @@
+export { default as NewComment } from "./NewComment";
