+++ src/app/views/components/ui/icons/DeleteIcon.tsx
