--- src/components/CardModal/components/NewComment/index.tsx
@@ -1,1 +1,0 @@
-export {default as NewComment} from './NewComment';
