+++ src/index.tsx
@@ -1,11 +1,14 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import App from "./App";
-import "./styles/normalize.scss";
-
+import "./app/views/styles/normalize.scss";
+import { Provider } from "react-redux";
+import { store } from "app/state/store";
 ReactDOM.render(
   <React.StrictMode>
+    <Provider store={store}>
       <App />
+    </Provider>
   </React.StrictMode>,
   document.getElementById("root")
 );
